1.3.9: Container fixes, toggle effects, improved efficiency
This is a "critical" update for Survival servers as I discovered an item duplication issue in 1.3.8 and prior versions of this plugin. This should be fixed in 1.3.9 and going forward.

Double chests will now restore contents correctly (this may not have been working previously) as long as multiple explosions are not simultaneously going off around the same double chest.

Additionally, you can now toggle off the particle and sound effects by setting these respective values to an empty string ("") in your config.

The only other noteworthy changes are efficiency improvements.