1.4.1: Item drops
This version adds support for 1.16.3 and a toggle to allow item drops from exploded blocks. This functionality was previously disabled due to potential item duplication issues, and while I do not recommend enabling item drops, you can now do so with the [icode]drops:[/icode] section in the config. The default blacklist has been provided by Linksu (thank you) - if drops are enabled, material names from this blacklist will still NOT be dropped on the ground.